Transitions adapted to full screen

+++ app/src/main/java/milespeele/canvas/util/TextUtils.java
@@ -43,35 +43,21 @@ public final class TextUtils {
         textPaint.setTextSize(100);
         textPaint.setTextScaleX(1.0f);
         Rect bounds = new Rect();
-        // ask the paint for the bounding rect if it were to draw this
-        // text
         textPaint.getTextBounds(text, 0, text.length(), bounds);
-        // get the height that would have been produced
         int h = bounds.height();
-        // make the text text up 70% of the height
         float target = (float) height * .2f;
-        // figure out what textSize setting would create that height
-        // of text
         float size = ((target/h) * 100f);
-        // and set it into the paint
         textPaint.setTextSize(size);
     }
 
     public static  void adjustTextScale(Paint textPaint, String text, float width, int paddingLeft,
                                          int paddingRight) {
-        // do calculation with scale of 1.0 (no scale)
         textPaint.setTextScaleX(1.0f);
+
         Rect bounds = new Rect();
-        // ask the paint for the bounding rect if it were to draw this
-        // text.
         textPaint.getTextBounds(text, 0, text.length(), bounds);
-        // determine the width
         int w = bounds.width();
-        // calculate the baseline to use so that the
-        // entire text is visible including the descenders
-        // determine how much to scale the width to fit the view
-        float xscale = ((float) (width - paddingLeft - paddingRight)) / w;
-        // set the scale for the text paint
+        float xscale = (width - paddingLeft - paddingRight) / w;
         textPaint.setTextScaleX(xscale * .6f);
     }
